+++ FinanceInvoiceCompare/FinanceInvoiceCompare.Model/ViewModels/UMatchInvoiceReportViewModel.cs
@@ -12,7 +12,7 @@ namespace FinanceInvoiceCompare.WebApi.Model
 
         public DateTime? InvoiceDate { get; set; }
 
-        public float? Amount { get; set; }
+        public decimal? Amount { get; set; }
 
         public DateTime? EntryDate { get; set; }
 
@@ -39,7 +39,7 @@ namespace FinanceInvoiceCompare.WebApi.Model
 
         public string Curr { get; set; }
 
-        public float? AmountInDC { get; set; }
+        public decimal? AmountInDC { get; set; }
 
         public string PBk { get; set; }
 
@@ -49,7 +49,7 @@ namespace FinanceInvoiceCompare.WebApi.Model
 
         public DateTime? BlineDate { get; set; }
 
-        public float? AmtLC2 { get; set; }
+        public decimal? AmtLC2 { get; set; }
 
         public string Assign { get; set; }
 
@@ -60,6 +60,6 @@ namespace FinanceInvoiceCompare.WebApi.Model
         public string VendorChName { get; set; }
 
 
-        public float? Check { get; set; }
+        public decimal? Check { get; set; }
     }
 }
